Update MIS API.

+++ atlas.com/marg/models/monster_spawn_point.go
@@ -1,14 +1,14 @@
 package models
 
 type MonsterSpawnPoint struct {
-	Id      int
-	MobTime int
-	Team    int
-	Cy      int
-	F       int
-	Fh      int
-	Rx0     int
-	Rx1     int
-	X       int
-	Y       int
+	Id      uint32
+	MobTime uint32
+	Team    int32
+	Cy      int16
+	F       uint32
+	Fh      uint16
+	Rx0     int16
+	Rx1     int16
+	X       int16
+	Y       int16
 }
